// client/desktop/www/src/app/workers/desktop/background/starlight/Shared.ts
/**
 * @license     BSL 1.1 (subject to the terms defined in the project's LICENSE file)
 * @copyright   (c) 2023, John Brandle
 */

export enum StarlightBackgroundTask
{
    init = 'init',
    fadeIn = 'fadeIn',
    fadeOut = 'fadeOut',
    add = 'add',
    performance = 'performance',
    size = 'size',
    page = 'page',
    beta = 'beta',
    shift = 'shift',
    rotate = 'rotate',
    visibilityState = 'visibilityState',
    focusState = 'blurState',
}